requirements.txt
flask==3.0.3
gunicorn==22.0.0

numpy==1.26.4
pillow==10.4.0

torch==2.2.2
torchvision==0.17.2

tensorflow==2.15.0